% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flags.R
\name{getITKcompileflags}
\alias{getITKcompileflags}
\title{return ITK compile flags}
\usage{
getITKcompileflags()
}
\description{
call this to get the compilation flags used for ITK and dependent software
}
\examples{
getITKcompileflags()
}
\author{
Avants BB
}

