% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flags.R
\name{getITKincludes}
\alias{getITKincludes}
\title{return ITK installation information}
\usage{
getITKincludes()
}
\description{
call this to get the include path for ITK headers and libraries
}
\examples{
getITKincludes()
}
\author{
Avants BB
}

