% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flags.R
\name{getITKversion}
\alias{getITKversion}
\title{return ITK version information}
\usage{
getITKversion()
}
\description{
call this to get the installed ITK version
}
\examples{
getITKversion()
}
\author{
Avants BB
}

