% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getRITKincludes.R
\name{getRITKincludes}
\alias{getRITKincludes}
\title{returns the include path}
\usage{
getRITKincludes()
}
\description{
call this to get the include path for ITK headers
}
\examples{
getRITKincludes()
}
\author{
Avants BB
}

