% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getRITKlibs.R
\name{getRITKlibs}
\alias{getRITKlibs}
\title{returns the libraries}
\usage{
getRITKlibs()
}
\description{
call this to get the ITK library locations
}
\examples{
getRITKlibs()
}
\author{
Avants BB
}

