% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{itkCompileFlags}
\alias{itkCompileFlags}
\title{return ITK compile flags}
\usage{
itkCompileFlags()
}
\description{
call this to get the compilation flags used for ITK and dependent software
}
\examples{

itkCompileFlags()

}
\author{
Avants BB
}
