% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{itkDir}
\alias{itkDir}
\title{return ITK directory information}
\usage{
itkDir()
}
\description{
call this to get the include path for ITK configuration.
this is what needs to be passed to dependent programs.
}
\examples{

itkDir()

}
\author{
Avants BB
}
