% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flags.R
\name{itkIncludes}
\alias{itkIncludes}
\title{return ITK installation information}
\usage{
itkIncludes()
}
\description{
call this to get the include path for ITK headers and libraries
}
\examples{
itkIncludes()
}
\author{
Avants BB
}

