% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flags.R
\name{itkLibs}
\alias{itkLibs}
\title{return ITK library location}
\usage{
itkLibs()
}
\description{
call this to get the path to ITK static libaries to which you will link
}
\examples{
itkLibs()
}
\author{
Avants BB
}

