% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flags.R
\name{itkVersion}
\alias{itkVersion}
\title{return ITK version information}
\usage{
itkVersion()
}
\description{
call this to get the installed ITK version
}
\examples{
itkVersion()
}
\author{
Avants BB
}

